fix: remove telemetry settings from runtime settings
Admins should not be able to see or control Segment telemetry
configuration; these were only exposed to simplify test setup. Removes
segment_write_key, segment_endpoint, and collection_interval_seconds
from the runtime settings catalog along with the live-reload plumbing
that watched them, leaving telemetry driven solely by static env vars.
