+++ backend/src/syntara/api/main.py
@@ -186,10 +186,6 @@ async def _lifespan_startup(app: FastAPI) -> dict[str, Any]:  # noqa: PLR0915
     # runtime override has been set by an operator).
     await apply_runtime_log_level()
 
-    # Import telemetry watcher so @watch_setting("telemetry.segment_write_key")
-    # is registered before start_watching() applies pending watchers.
-    import syntara.telemetry.client  # noqa: F401, PLC0415
-
     # Watch for runtime log level changes and start polling
     runtime_settings.start_watching()
 
@@ -246,7 +242,6 @@ async def _lifespan_startup(app: FastAPI) -> dict[str, Any]:  # noqa: PLR0915
     # Initialize periodic analytics collector
     periodic_collector = PeriodicCollector(
         registry=get_telemetry_registry(),
-        settings_cache=runtime_settings,
     )
 
     completion_poller = get_completion_poller()
