+++ backend/tests/unit/telemetry/test_client.py
@@ -117,104 +117,3 @@ class TestTelemetryClientRegistry:
         with patch("syntara.telemetry.client.logger") as mock_logger:
             TelemetryClientRegistry._error_handler(RuntimeError("test"), [{"item": 1}])
             mock_logger.warning.assert_called_once()
-
-
-class TestTelemetryClientRegistryReinitialize:
-    """Tests for TelemetryClientRegistry.reinitialize()."""
-
-    @patch("syntara.telemetry.client.get_settings")
-    @patch("syntara.telemetry.client.segment_analytics")
-    def test_reinitialize_creates_new_client(self, mock_segment, mock_get_settings):
-        mock_settings = MagicMock()
-        mock_settings.segment_endpoint = "https://api.segment.io"
-        mock_settings.segment_max_retries = 10
-        mock_settings.segment_timeout = 30
-        mock_get_settings.return_value = mock_settings
-
-        old_client = MagicMock()
-        new_client = MagicMock()
-        mock_segment.Client.side_effect = [old_client, new_client]
-
-        registry = TelemetryClientRegistry()
-        registry.initialize(write_key="old-key")
-        assert registry.is_initialized()
-
-        registry.reinitialize("new-key")
-
-        old_client.flush.assert_called_once()
-        assert mock_segment.Client.call_count == 2
-        assert registry.is_initialized()
-
-    @patch("syntara.telemetry.client.segment_analytics")
-    def test_reinitialize_empty_key_disables(self, mock_segment):
-        mock_client = MagicMock()
-        mock_segment.Client.return_value = mock_client
-
-        registry = TelemetryClientRegistry()
-        registry.initialize(write_key="old-key")
-
-        registry.reinitialize("")
-
-        mock_client.flush.assert_called_once()
-        assert registry.is_initialized() is False
-
-    def test_reinitialize_when_not_initialized(self):
-        registry = TelemetryClientRegistry()
-        registry.reinitialize("")
-        assert registry.is_initialized() is False
-
-    @patch("syntara.telemetry.client.get_settings")
-    @patch("syntara.telemetry.client.segment_analytics")
-    def test_reinitialize_uses_custom_host(self, mock_segment, mock_get_settings):
-        mock_settings = MagicMock()
-        mock_settings.segment_endpoint = "https://api.segment.io"
-        mock_settings.segment_max_retries = 10
-        mock_settings.segment_timeout = 30
-        mock_get_settings.return_value = mock_settings
-
-        mock_segment.Client.return_value = MagicMock()
-
-        registry = TelemetryClientRegistry()
-        registry.reinitialize("test-key", host="http://mock-segment:9999")
-
-        call_kwargs = mock_segment.Client.call_args[1]
-        assert call_kwargs["host"] == "http://mock-segment:9999"
-        assert call_kwargs["write_key"] == "test-key"
-
-    @patch("syntara.telemetry.client.get_settings")
-    @patch("syntara.telemetry.client.segment_analytics")
-    def test_reinitialize_falls_back_to_static_host(self, mock_segment, mock_get_settings):
-        mock_settings = MagicMock()
-        mock_settings.segment_endpoint = "https://custom.segment.io"
-        mock_settings.segment_max_retries = 10
-        mock_settings.segment_timeout = 30
-        mock_get_settings.return_value = mock_settings
-
-        mock_segment.Client.return_value = MagicMock()
-
-        registry = TelemetryClientRegistry()
-        registry.reinitialize("test-key")
-
-        call_kwargs = mock_segment.Client.call_args[1]
-        assert call_kwargs["host"] == "https://custom.segment.io"
-
-    @patch("syntara.telemetry.client.get_settings")
-    @patch("syntara.telemetry.client.segment_analytics")
-    def test_reinitialize_flushes_before_replacing(self, mock_segment, mock_get_settings):
-        mock_settings = MagicMock()
-        mock_settings.segment_endpoint = "https://api.segment.io"
-        mock_settings.segment_max_retries = 10
-        mock_settings.segment_timeout = 30
-        mock_get_settings.return_value = mock_settings
-
-        old_client = MagicMock()
-        new_client = MagicMock()
-        mock_segment.Client.side_effect = [old_client, new_client]
-
-        registry = TelemetryClientRegistry()
-        registry.initialize(write_key="old-key")
-
-        registry.reinitialize("new-key")
-
-        old_client.flush.assert_called_once()
-        new_client.flush.assert_not_called()
